ajout d'une classe pour calculer les notes finaux et remplacement de integer en double

+++ src/main/java/org/example/History.java
@@ -7,11 +7,11 @@ import java.time.Instant;
 
 @Getter
 public class History {
-    private final Integer note;
+    private final Double note;
     private final Instant time;
     private final String reason;
 
-    public History(Integer note, Instant time, String reason) {
+    public History(Double note, Instant time, String reason) {
         this.note = note;
         this.time = time;
         this.reason = reason;
